% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-capuchin-data.R
\name{calc_dsi_topn}
\alias{calc_dsi_topn}
\title{Averages the DSI over the top n partners}
\usage{
calc_dsi_topn(
  dsi_table,
  slots = 3,
  group_true = FALSE,
  year_true = FALSE,
  dsi_column = "DSI"
)
}
\arguments{
\item{dsi_table}{Table with DSI values for dyads of IndA and IndB}

\item{slots}{Set the number of top partners to be included}

\item{group_true}{Set to TRUE, if the DSI should be averaged per group.
Default is FALSE in case no GroupCode-column is provided. Make sure that
column GroupCode exists when set to TRUE.}

\item{year_true}{Set to TRUE, if the DSI should be averaged per year. Default
is FALSE in case no Year-column is provided (I also believe this is what
Silk et al. did in their publications). Make sure that column YearOf exists
when set to TRUE.}

\item{dsi_column}{Specify how the DSI column is labelled. Default is "weight"
as for the Social Network Parameter calculation}
}
\description{
Averages for entire study period over all groups or per group and per year if specified
}
\examples{


}
