% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-capuchin-data.R
\name{calc_dyadic_rates}
\alias{calc_dyadic_rates}
\title{Calculation of behavior rates for female dyads}
\usage{
calc_dyadic_rates(focaldata, census_pace, biography_pace, group, start, end,
  minAgeF, behaviours_duration, behaviours_events, minEffortDyad = 0,
  minEffortInd = 0, roles_included = c("D", "R", "M"),
  dyads_summarized = TRUE)
}
\arguments{
\item{focaldata}{Dataset with focaldata from capuchin monkeys}

\item{census_pace}{Dataframe from the PACE database with monthly capuchin census data}

\item{biography_pace}{Dataframe from the PACE database with individual biography data}

\item{group}{Name of the social group}

\item{start}{Define first and last day (YYYY-MM-DD) of the desired time period}

\item{end}{Define first and last day (YYYY-MM-DD) of the desired time period}

\item{minAgeF}{The minimum age of females to be included (in years)}

\item{behaviours_duration}{Behaviours with duration (e.g. grooming) to be
included}

\item{minEffortDyad}{Minimum duration (in hours) a dyad has to be observed to
be included (observation time of IndA while IndB was around + observation
time IndB while IndA was around)}

\item{minEffortInd}{The minimum duration (in hours) an individual has to be
observed to be included}

\item{roles_included}{Which roles should be included, e.g., (D)irected,
(R)eceived, or (M)utual}

\item{dyads_summarized}{Should dyadic values be summarized? If FALSE, each
behavior will be listed as IndA->IndB and as IndB->IndA separately (per
dyad and per role)}

\item{beahviours_events}{Behaviours without duration (e.g. approach) to be
included}
}
\value{
If values are summarized, the rates IndA->IndB and IndB->IndA are
  averaged. This is slightly different from calculating behaviour counts (or
  durations) IndA->IndB + IndB->IndA divided by EffortA + EffortB (unless
  EffortA = EffortB). Thus, rates are not weighted by focal effort of IndA
  and IndB
}
\description{
This function calculates rates of different behaviors for all female dyads
meeting specific criteria (age, focal effort) within a specified group and
time period (only works with one group and one time period at a time).
}
