% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-capuchin-data.R
\name{calc_dyadic_rates_mymg}
\alias{calc_dyadic_rates_mymg}
\title{Calculate dyadic rates of interactions for several groups and years}
\usage{
calc_dyadic_rates_mymg(focaldata, census_pace, biography_pace, groups,
  years, minAgeF, behaviours_duration, behaviours_events, minEffortDyad,
  minEffortInd, roles_included = c("D", "R", "M"),
  dyads_summarized = TRUE)
}
\arguments{
\item{focaldata}{Dataset with focaldata from capuchin monkeys}

\item{census_pace}{Dataframe from the PACE database with monthly capuchin census data}

\item{biography_pace}{Dataframe from the PACE database with individual biography data}

\item{groups}{Vector with all social groups to be included}

\item{years}{Vector will all years to be included}

\item{minAgeF}{The minimum age of females to be included (in years)}

\item{behaviours_duration}{Behaviours with duration (e.g. grooming) to be
included}

\item{minEffortDyad}{Minimum duration (in hours) a dyad has to be observed to
be included (observation time of IndA while IndB was around + observation
time IndB while IndA was around)}

\item{minEffortInd}{The minimum duration (in hours) an individual has to be
observed to be included}

\item{roles_included}{Which roles should be included, e.g., (D)irected,
(R)eceived, or (M)utual}

\item{dyads_summarized}{Should dyadic values be summarized? If FALSE, each
behavior will be listed as IndA->IndB and as IndB->IndA separately (per
dyad and per role)}
}
\description{
Based on \code{calc_dyadic_rates} but works for several groups and years.
Thus, function calculates rates of interactions for all female dyads meeting
specific criteria (age, focal effort) within each of the specified groups and
years.
}
\examples{

}
