% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-capuchin-data.R
\name{calc_effort_AB}
\alias{calc_effort_AB}
\title{Calculate focal time for female A while female B was present}
\usage{
calc_effort_AB(focaldata, census_pace, biography_pace, group, start, end,
  minAgeF)
}
\arguments{
\item{focaldata}{Dataset with focal data from capuchin monkeys}

\item{census_pace}{Dataframe from the PACE database with monthly capuchin census data}

\item{biography_pace}{Dataframe from the PACE database with individual biography data}

\item{group}{The name of the social group for which the presence of females
will be determined}

\item{start}{Define first and last day (YYYY-MM-DD) of the time period
for which the presence of females will be determined}

\item{end}{Define first and last day (YYYY-MM-DD) of the time period
for which the presence of females will be determined}

\item{minAgeF}{The minimum age of females to be included (in years)}
}
\description{
Calculates focal effort with female A while female B was present for the
specifiec group and time period
}
\examples{
effort_AB(focaldata = dataset, census_pace = Vcensus, biography_pace =
Vbiography, group = "GUAN", start = "2007-01-01", end = "2011-12-31", minAgeF
= 5)
}
