% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-redcolobus-data.R
\name{calc_ind_NN_together}
\alias{calc_ind_NN_together}
\title{Function to calculate for all individuals, for each month the number of lines
(i.e. observations) where A was Individual with B as NN}
\usage{
calc_ind_NN_together(
  ind_nn_single_timeperiod,
  dyad_table,
  dyad_matrix,
  max_nn_dist,
  return_matrix = TRUE
)
}
\arguments{
\item{ind_nn_single_timeperiod}{based on ind_nn_df created with
`prep_red_colobus_ind_nn_data()`}

\item{dyad_table}{Data frame with all individuals observed during the time
period indicating for each month whether they were adult}

\item{dyad_matrix}{Empty matrix with all individual names as rows/cols}

\item{max_nn_dist}{The maximum distance for which an individual is still
considered NN. Default is 5 (m)}

\item{return_matrix}{Should the result returned as matrix (default) or data
frame}
}
\description{
Function to calculate for all individuals, for each month the number of lines
(i.e. observations) where A was Individual with B as NN
}
\examples{


}
