% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-sna-pars.R
\name{calc_LL_S}
\alias{calc_LL_S}
\title{Function to calculate likelihood of mu and S (for `get_S``)}
\usage{
calc_LL_S(par, data)
}
\arguments{
\item{par}{vector with the two parameters mu and S}

\item{data}{Data frame with columns dvec and xvec}
}
\description{
For details, see `get_S`
}
