% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calc_matrix_correlation}
\alias{calc_matrix_correlation}
\title{Calculates correlation between two (symmetrical) matrices using `stats::cor`}
\usage{
calc_matrix_correlation(
  m1,
  m2,
  zeros_remove = FALSE,
  method = c("pearson", "kendall", "spearman")
)
}
\arguments{
\item{m1, m2}{The two matrices. Function only keeps row and colnames that exist in m1 and m2.}

\item{zeros_remove}{Should all zeros be removed? (set to NA before running `cor`).}

\item{method}{a character string indicating which correlation
    coefficient (or covariance) is to be computed.  One of
    \code{"pearson"} (default), \code{"kendall"}, or \code{"spearman"}:
    can be abbreviated.}
}
\description{
Calculates correlation between two (symmetrical) matrices using `stats::cor`
}
