% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-sna-pars.R
\name{calc_ppi}
\alias{calc_ppi}
\title{Calculates the Partner Preference Index (PPI) based on two network matrices.}
\usage{
calc_ppi(m1, m2, n_Slots = 3, details = FALSE)
}
\arguments{
\item{m1, m2}{The two matrices.}

\item{n_Slots}{number of top partners evaluated (S in Silk et al. 2013)}

\item{if}{FALSE, returns a single value. If TRUE, a dataframe with details}
}
\description{
This function calculate the Partner Preference Index (PPI) as described by
Silk, Cheney and Seyfarth (2013, Evolutionary Anthropology):
}
\details{
PPI = (2S -  U)/(2S - S - X) with S = number of partner rank slots being
evaluated (here `n_Slots`), U = Number of different top partners that the
individual had in those years, and X = Number of top partners present in t1,
but not t2:
}
