% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-sna-pars.R
\name{calc_sna_pars_dyad}
\alias{calc_sna_pars_dyad}
\title{Calculate various individual social network parameters based on DSI values}
\usage{
calc_sna_pars_dyad(edge_table, indA_col = "IndA", indB_col = "IndB",
  weight_col = "DSI")
}
\arguments{
\item{edge_table}{Table with two individual and one weight column}

\item{indA_col, indB_col, weight_col}{Specify names of columns for the two
individuals and the weight to be used in the SNA}
}
\description{
Determines Strength, Betweenness, Eigenvector centrality, Clustering
coefficient (cc), Reach. Works with weighted, but undirected edges
}
\examples{

}
