% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-sna-pars.R
\name{calc_sna_pars_mymg}
\alias{calc_sna_pars_mymg}
\title{Calculate various individual social network parameters based on DSI values
for multiple years and groups}
\usage{
calc_sna_pars_mymg(edge_table, indA_col = "IndA", indB_col = "IndB",
  weight_col = "DSI")
}
\arguments{
\item{edge_table}{Table with two individual and one weight column. Also
requires columns "GroupCode" and "YearOf" or will fail.}

\item{indA_col}{Specify names of columns for the two
individuals and the weight to be used in the SNA}

\item{indB_col}{Specify names of columns for the two
individuals and the weight to be used in the SNA}

\item{weight_col}{Specify names of columns for the two
individuals and the weight to be used in the SNA}
}
\description{
Determines Strength, Betweenness, Eigenvector centrality, Clustering
coefficient (cc), Reach. Works with weighted, but undirected edges. Works
with several groups and years
}
\examples{

}
