% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-redcolobus-data.R
\name{create_age_sex_codes}
\alias{create_age_sex_codes}
\title{Red Colobus: Create age-sex codes based on the link table}
\usage{
create_age_sex_codes(link_table)
}
\arguments{
\item{link_table}{Table creates with \code{create_link_table}}
}
\description{
Red Colobus: Create age-sex codes based on the link table
}
\examples{


}
