% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-redcolobus-data.R
\name{create_link_table}
\alias{create_link_table}
\title{Red Colobus: Create table to link all synonyms to an individual name, and
also includes the sex of individuals}
\usage{
create_link_table(individual_table)
}
\arguments{
\item{individual_table}{IndividualsRC table should include all individuals
with all possible alternative names for each individual}
}
\description{
This function uses the IndividualsRC table to create a longer table with
names, link names (synonyms) and sex of individuals. It also included unknown
individuals (e.g. Adult Female) with the code number as synonym, and the sex
(if known)
}
\examples{


}
