% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-capuchin-data.R
\name{determine_presence}
\alias{determine_presence}
\title{Determine presence of female in group}
\usage{
determine_presence(census_pace, biography_pace, group, start, end, minAgeF)
}
\arguments{
\item{census_pace}{Dataframe from the PACE database with monthly capuchin census data}

\item{biography_pace}{Dataframe from the PACE database with individual biography data}

\item{group}{The name of the social group for which the presence of females
will be determined}

\item{start, end}{Define first and last day (YYYY-MM-DD) of the time period
for which the presence of females will be determined}

\item{minAgeF}{The minimum age of females to be included (in years)}
}
\description{
This function uses census and biography data from the PACE database to
determine the periods of time females with specified minimum age
\code{minAgeF} were present in a \code{group} from \code{start} to
\code{end}.
The function does only work for females with known DOB
}
\examples{
female_presence(census_pace, biogaphy_pace, group = "GUAN", start =
"2007-01-01", end = "2011-12-31", minAgeF = 5)
}
