% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-redcolobus-data.R
\name{determine_rc_presence}
\alias{determine_rc_presence}
\title{Determine presence of adult individuals in the Red colobus group for the specified period of time}
\usage{
determine_rc_presence(scandata, first_month_adult_df, from_date, to_date,
  n_min = 5, details = FALSE)
}
\arguments{
\item{scandata}{Scan data frame}

\item{first_month_adult_df}{Dataframe with first-month_adult}

\item{from_date, to_date}{Start and end date of analyzed period of time. Format should be "yyyy-mm-dd"}

\item{n_min}{Min. number of times and individual has to be recorded (Individual + NN) to be considered present}

\item{details}{Provide details for each individual and each month?}
}
\description{
This function uses scan data in combination with data on the first-adult-month of individual
to determine for each month whether an individual was adult and present
}
\details{
Default is that individuals are considered absent if recorded less than \code{n_min} (default 5) times per month for 3 months in a row
}
\examples{


}
