% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-capuchin-data.R
\name{dyadic_rates_to_dsi}
\alias{dyadic_rates_to_dsi}
\title{Function calculates DSI from dyadic rates}
\usage{
dyadic_rates_to_dsi(
  dyadic_rates_table,
  behaviours_included,
  roles_included = c("D", "R", "M")
)
}
\arguments{
\item{dyadic_rates_table}{Table that was created by \code{calc_dyadic_rates}}

\item{behaviours_included}{Vector with all behaviours to be included}

\item{roles_included}{Which roles should be included, e.g., (D)irected,
(R)eceived, or (M)utual}
}
\description{
This functions uses the table with dyadic rates calculated by
\code{calc_dyadic_rates} to calculate the Dyadic Sociality Index (DSI) for
all dyads included in the dataset. Included behaviours and roles can be
specified (as in \code{calc_dyadic_rates})
}
\examples{

}
