% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-capuchin-data.R
\name{dyadic_rates_to_dsi_mymg}
\alias{dyadic_rates_to_dsi_mymg}
\title{Function calculates DSI from dyadic rates for several groups and years}
\usage{
dyadic_rates_to_dsi_mymg(
  dyadic_rates_table_mymg,
  behaviours_included,
  roles_included = c("D", "R", "M")
)
}
\arguments{
\item{dyadic_rates_table_mymg}{Table that was created by
\code{calc_dyadic_rates_mymg}}

\item{behaviours_included}{Vector with all behaviours to be included}

\item{roles_included}{Which roles should be included, e.g., (D)irected,
(R)eceived, or (M)utual}
}
\description{
As \code{dyadic_rates_to_dsi}, this function calculates the Dyadic Sociality
Index (DSI) for all dyads included in the dataset based on dyadic interaction
rates. However, this function uses the table with dyadic rates from multiple
groups and years calculated by \code{calc_dyadic_rates_mymg} instead of the
table with a single group from a single year.
}
\examples{

}
