% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-sna-pars.R
\name{get_CCs_from_timeperiod_list}
\alias{get_CCs_from_timeperiod_list}
\title{Function to calculate the global Clusterin Coefficient (CC) from simple ratios.}
\usage{
get_CCs_from_timeperiod_list(simple_ratio_list, perm.by.layer = NULL)
}
\arguments{
\item{simple_ratio_list}{List created with `get_simple_ratios()``}

\item{perm.by.layer}{In case not all layers (matrices) of permuted data
should be included. This is the `by` argument of the `seq()` function.}
}
\description{
This function uses a list of the Simple Ratio Association Index (or any other
index) for all included time periods. It then calculate the CC for the
observed data and for matrices of permuted social networks
}
