% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-sna-pars.R
\name{get_cluster_estimates}
\alias{get_cluster_estimates}
\title{Function to estimates of clustering in a network using different algorithms}
\usage{
get_cluster_estimates(m)
}
\arguments{
\item{m}{symmetric adjajency matrix of a network with values reflecting
relationships among individuals.}
}
\description{
This function uses a symmetric adjacancy matrix and uses the following
functions from the `igraph` package to estimate the number of clusters and
the modularity in a network: `cluster_fast_greedy`, `cluster_infomap`,
`cluster_louvain`, `cluster_optimal`, `cluster_walktrap`
}
