% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-sna-pars.R
\name{get_cluster_info_from_timeperiod_list}
\alias{get_cluster_info_from_timeperiod_list}
\title{Function that uses two algorithms (fast_greedy, infomap) to get info about
clusters from simple ratios.}
\usage{
get_cluster_info_from_timeperiod_list(
  simple_ratio_list,
  perm.by.layer = NULL,
  cluster_methods = NULL
)
}
\arguments{
\item{simple_ratio_list}{List created with `get_simple_ratios()``}

\item{perm.by.layer}{In case not all layers (matrices) of permuted data
should be included. This is the `by` argument of the `seq()` function.}

\item{cluster_methods}{one or more of the `igraph::cluster_` function (see `?igraph::communities` for available functions).}
}
\description{
This function uses a list of the Simple Ratio Association Index (or any other
index) for all included time periods. It then calculates modularity and the
ratio of individuals/estimated groups scaled from 0 to 1
(ratio_clusters_inds_scaled) from the observed data and for matrices of permuted
social networks. It use two algorithms from the `igraph` package:
cluster_infomap and cluster_fast_greedy.
Publications that discuss different ways of assessing modularity:
Scott Emmons et al, 2016, Plos One (infomap, Blondel, label propagation, smart local moving)
}
