% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-sna-pars.R
\name{get_communities}
\alias{get_communities}
\title{Function to estimates of clustering in a network using the specified algorithm.}
\usage{
get_communities(m, cluster_method = NULL)
}
\arguments{
\item{m}{symmetric adjajency matrix of a network with values reflecting relationships among individuals.}

\item{cluster_method}{see `?igraph::communities` for available functions. E.g. `cluster_fastgreedy`}
}
\description{
This function uses a symmetric adjacancy matrix and uses one of the `igraph::cluster_`
functions to estimate the number of clusters and the modularity of a network.
}
