% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-sna-pars.R
\name{get_CVs_from_timeperiod_list}
\alias{get_CVs_from_timeperiod_list}
\title{Function to calculate Coefficients of variation (CV) of association indices
Whitehead (2008) suggests to use his estimate of S, instead, but this
approach is also commonly used.}
\usage{
get_CVs_from_timeperiod_list(
  simple_ratio_list,
  perm.by.layer = NULL,
  zeros.rm = FALSE,
  na.rm = FALSE
)
}
\arguments{
\item{simple_ratio_list}{List created with `get_simple_ratios()``}

\item{perm.by.layer}{In case not all layers (matrices) of permuted data
should be included. This is the `by` argument of the `seq()` function.}

\item{zeros.rm}{Should all zeros be removed (i.e. dyads that were never
observed together). Could be problematic with permuted networks.}

\item{na.rm}{Should NAs be removed. This argument is probably redudant as all
NAs are set to 0 (which is correct in this case). But if set to FALSE, it
controls if all matrices were well constructed.}
}
\description{
This function uses a list of the Simple Ratio Association Index (or any other
index) for all included time periods. It then calculate the CV for the
observed data and for matrices of permuted social networks
}
