% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-redcolobus-data.R
\name{get_dyadic_association_rc}
\alias{get_dyadic_association_rc}
\title{Count number of scans for a period with Individual while it is present and adult
Then count for each other Individual number of times it was Nearest Neighbour (NN) or could have
been NN while adult and present in group}
\usage{
get_dyadic_association_rc(scan_data_df, first_month_adult_df, id_sex_df,
  from_date, to_date, max_nn_dist = 5)
}
\arguments{
\item{first_month_adult_df}{Dataframe with first-month-adult dates}

\item{id_sex_df}{Table with sex of individuals. This is the link table creates with `create_link_table`}

\item{from_date, to_date}{Start and end date of analyzed period of time.
Format should be "yyyy-mm-dd"}

\item{max_nn_dist}{The maximum distance for which an individual is still considered NN. Default is 5}

\item{scandata_df}{Data frame with red colobus scan data}
}
\description{
Count number of scans for a period with Individual while it is present and adult
Then count for each other Individual number of times it was Nearest Neighbour (NN) or could have
been NN while adult and present in group
}
\examples{


}
