% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-redcolobus-data.R
\name{get_dyadic_observations}
\alias{get_dyadic_observations}
\title{For each dyad within each time period, this function counts the number of
times an Individual A was recorded as Individual while Individual A and B
were both adult and present (`$n_Ind_while_NN_present`). Then, it calculates
how often Individual B was recorded as the Nearest Neighbour (NN) of
Individual A for a time period (where both were adult and present;
`$n_Ind_NN_together_observed`).}
\usage{
get_dyadic_observations(
  ind_nn_df,
  id_sex_df,
  presence_df,
  max_nn_dist = 5,
  n_permutations = 0,
  set_seed = 1209,
  ind_sex_permutation = c("Male", "Female"),
  nn_sex_permutation = c("Male", "Female"),
  flips_per_permutation = 1,
  ...
)
}
\arguments{
\item{ind_nn_df}{Data frame with information on indviduals and nearest
neighours from red colobus scan data. Created with
`prep_red_colobus_ind_nn_data()`}

\item{id_sex_df}{Table with sex of individuals. This is the link table
created with `create_link_table`}

\item{presence_df}{Dataframe created with `get_presence_rc`}

\item{max_nn_dist}{The maximum distance for which an individual is still
considered NN. Default is 5 (m)}

\item{n_permutations}{How many permutations should be conducted?}

\item{ind_sex_permutation}{default `c("Male", "Female")`}

\item{nn_sex_permutation}{default `c("Male", "Female")`}

\item{flips_per_permutation}{Number of iterations ('NN flips') per
permutation}
}
\description{
Furthermore, this function conducts pre-network data permutations
`n_permutations` times (0 for no permutations) using the function
`permute_scan_data` For each permutation, `flips_per_permutation` NNs are
flipped. Then, the number of times Individual B was recorded as NN of
Individual A is counted (`$n_Ind_NN_together_permuted`)
}
\details{
Limitations: This function cannot (yet) permute within sex classes only. To
achieve that, subset scandata and only include females or males before
running the function. Set `ind_sex_permutation` and `nn_sex_permutation` to
"Male"/"Female" respectively (although this should be working with default
arguments)
}
\examples{

}
