% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-sna-pars.R
\name{get_fast_greedy}
\alias{get_fast_greedy}
\title{Function to estimates of clustering in a network using the fast_greedy algorithm.}
\usage{
get_fast_greedy(m)
}
\arguments{
\item{m}{symmetric adjajency matrix of a network with values reflecting relationships among individuals.}
}
\description{
This function uses a symmetric adjacancy matrix and uses the `igraph::cluster_fast_greedy`
function to estimate the number of clusters and the modularity of a network.
}
