% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-sna-pars.R
\name{get_matrix_cors_from_timeperiod_list}
\alias{get_matrix_cors_from_timeperiod_list}
\title{Get matrix correlations from timeperiod list}
\usage{
get_matrix_cors_from_timeperiod_list(
  simple_ratio_list,
  zeros_remove = FALSE,
  method = c("pearson", "kendall", "spearman")
)
}
\arguments{
\item{simple_ratio_list}{List created with `get_simple_ratios()`}

\item{zeros_remove}{Should all zeros be removed? (set to NA before running `cor`).}

\item{method}{a character string indicating which correlation
    coefficient (or covariance) is to be computed.  One of
    \code{"pearson"} (default), \code{"kendall"}, or \code{"spearman"}:
    can be abbreviated.}
}
\description{
Function to calculate correlation coefficients of all two consecutive
networks within a list of timeperiods, each with observed and permuted
association indices.
}
