% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-sna-pars.R
\name{get_ppi_from_timeperiod_list}
\alias{get_ppi_from_timeperiod_list}
\title{Get the Partner Preference Index (PPI) from timeperiod list with association indices.}
\usage{
get_ppi_from_timeperiod_list(simple_ratio_list, n_slots = 3)
}
\arguments{
\item{simple_ratio_list}{List created with `get_simple_ratios()`}
}
\description{
Function to calculate the Partner Preference Index (PPI) of all two consecutive
networks within a list of timeperiods, each with observed and permuted
association indices.
}
