% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-sna-pars.R
\name{get_S_from_array}
\alias{get_S_from_array}
\title{Calculate multiple estimates of Social Differentiation according to Whitehead
2008 from an array.}
\usage{
get_S_from_array(
  i_j_sum,
  i_j_together_array,
  initial.values = c(0.5, 0.5),
  lower = c(0.01, 0.01),
  upper = c(1, 10),
  use.parallel = FALSE,
  cores = 1
)
}
\arguments{
\item{i_j_sum}{Matrix with values i_j_sum[ij] used as denominator of the estimated association index (d in the Whitehead paper)}

\item{i_j_together_array}{Array with matrices with values i_j_together[ij] number of
observations of individuals i and j together (x in the Whitehead paper)}

\item{initial.values}{Initial values for the parameters to be optimized over (`par` argument in `optim``)}

\item{lower}{lower bounds for parameters}

\item{upper}{upper bounds for parameters}

\item{use.parallel}{TRUE/FALSE}

\item{cores}{Number of cores to be used for parallization}
}
\description{
For details, see `get_S_from_matrix()`. Difference is that this functions
takes an 3d-array, where each slice is an matrix of observed association,
e.g. for permuted networks. Can use parallization, which is highly
recommended.
}
\examples{

}
