% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-sna-pars.R
\name{get_S_from_matrix}
\alias{get_S_from_matrix}
\title{Calculate Estimate of Social Different according to Whitehead 2008}
\usage{
get_S_from_matrix(
  i_j_sum,
  i_j_together,
  initial.values = c(0.5, 0.5),
  lower = c(0.01, 0.01),
  upper = c(1, 10)
)
}
\arguments{
\item{i_j_sum}{Matrix with values i_j_sum[ij] used as denominator of the estimated association index (d in the Whitehead paper)}

\item{i_j_together}{Matrix with values i_j_together[ij] number of observations of individuals i and j together (x in the Whitehead paper)}

\item{initial.values}{Initial values for the parameters to be optimized over (`par` argument in `optim``)}

\item{lower}{lower bounds for parameters}

\item{upper}{upper bounds for parameters}
}
\description{
This function uses a max. likelhood to separate the variance of true association indices and the sampling variance.
It's described in "Precision and power in the analysis of social structure using associations" (Whitehead, 2008, Animal Behaviour) and
"Whitehead, H. 2008. Analyzing Animal Societies: Quantitative Methods for Vertebrate Social Analysis. Chicago: University of Chicago Press."
}
\examples{

}
