% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-sna-pars.R
\name{get_S_from_timeperiod_list}
\alias{get_S_from_timeperiod_list}
\title{Calculate multiple estimates of Social Differentiation according to Whitehead
2008 from a list of time periods, each with observed and permuted data.}
\usage{
get_S_from_timeperiod_list(
  dyadic_obs_list,
  perm.start.layer = 1,
  perm.by.layer = NULL,
  initial.values,
  lower,
  upper,
  use.parallel = FALSE,
  cores = 1
)
}
\arguments{
\item{dyadic_obs_list}{List created by `get_dyadic_observations()`}

\item{perm.start.layer}{Integer specifying the first permutation to be
included}

\item{perm.by.layer}{Integer specifying the the permutations to be inlcluded
(`by` argument in the `seq()` function)}

\item{initial.values}{Initial values for the parameters to be optimized over (`par` argument in `optim``)}

\item{lower}{lower bounds for parameters}

\item{upper}{upper bounds for parameters}

\item{use.parallel}{TRUE/FALSE}

\item{cores}{Number of cores to be used for parallization}
}
\description{
For details, see `get_S_from_matrix()` and `get_S_from_array()`. Difference
is that this functions takes a list of all time periods, each with observed
(in a matrix) and permuted data (in a 3d-array).
}
\examples{

}
