% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_index_funs.R
\name{get_simple_ratios}
\alias{get_simple_ratios}
\title{Function to calculate the simple ratio association index for several time period.}
\usage{
get_simple_ratios(
  dyadic_obs_list,
  i_j_observed_sum_min,
  new_dim_names = c("Name1", "Name2")
)
}
\arguments{
\item{dyadic_obs_list}{List created with `get_dyadic_observations`}

\item{i_j_observed_sum_min}{Minimum number of times individual i and j were observed while both adult and present}

\item{new_dim_names}{A vector with two elements for column and row names (default is Name1, Name2)}
}
\description{
So far, only works with Red Colobus Nearest Neighbour data
}
\examples{


}
