% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-sna-pars.R
\name{get_strength_from_rates}
\alias{get_strength_from_rates}
\title{Calculate individual strength}
\usage{
get_strength_from_rates(rate_table)
}
\arguments{
\item{rate_table}{Table with two individual ("IndA", "IndB") and one "Rate" column. Each dyad
should be listed twice (once in each direction)}
}
\description{
Determines Strength
}
\examples{

}
