% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-sna-pars.R
\name{get_strength_from_rates_mymg}
\alias{get_strength_from_rates_mymg}
\title{Calculate individual strength for multiple years and groups}
\usage{
get_strength_from_rates_mymg(rate_table)
}
\arguments{
\item{rate_table}{Table with two individual ("IndA", "IndB") and one "Rate" column. Each dyad
should be listed twice (once in each direction) per year. Also requires columns "GroupCode" and "YearOf" or will fail.}
}
\description{
Determines Strength
Works with several groups and years
}
\examples{

}
