% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-sna-pars.R
\name{get_transitivity_from_ratio_matrix}
\alias{get_transitivity_from_ratio_matrix}
\title{Calculates the global clustering coefficient (or transitivity) of a network.}
\usage{
get_transitivity_from_ratio_matrix(
  m,
  weighted = TRUE,
  diag = FALSE,
  mode = "lower"
)
}
\arguments{
\item{m}{Adjacency matrix with weights of connections.}
}
\description{
This function uses the `igraph::transitivity()` function to calculate the
global clustering coefficient of a network. The input should be a matrix with
something like an association index as the weight of connections between
nodes.
}
