% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{list_to_quantiles}
\alias{list_to_quantiles}
\title{Uses a named list (or vector of values) to calculate the quantiles of
specified probabilities}
\usage{
list_to_quantiles(value_list, probs = c(0.01, 0.98))
}
\arguments{
\item{value_list}{numeric vector or named list to calculate quantiles.}

\item{probs}{As specified in `quantile`}
}
\description{
If `value_list` is a vector, the function creates a data frame (tibble) with
two columns ("probability" and "quantile_value") and one row. If `value_list`
is a named list, the function creates a data frame with three columns
("group_var", "probability", and "quantile_value"), and one row per
list-element
}
