% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-redcolobus-data.R
\name{permute_scan_data}
\alias{permute_scan_data}
\title{Permutation of Nearest Neighbours in Red Colobus scan data: Following
suggestion by Farine (2017, MEE), this function permutes raw observations
('pre-network) data. This method has the advantage that "account for
underlying structure in the generated social network, and thus can reduce
both type I and type II error rates." (Farine, 2017)}
\usage{
permute_scan_data(
  ind_nn_single_timeperiod,
  presence_df,
  ind_sex_permutation = c("Male", "Female"),
  nn_sex_permutation = c("Male", "Female"),
  flips_per_permutation,
  show_i = FALSE
)
}
\arguments{
\item{ind_nn_single_timeperiod}{based on ind_nn_df created with
`prep_red_colobus_ind_nn_data()`}

\item{presence_df}{Dataframe created with `get_presence_rc`}

\item{ind_sex_permutation}{default `c("Male", "Female")`}

\item{nn_sex_permutation}{default `c("Male", "Female")`}

\item{flips_per_permutation}{Number of iterations ('NN flips') per
permutation}

\item{show_i}{Show progress}
}
\description{
It does so by choosing a random line from the scan data, and then checks if
Individual (1) and NN (1) are both adult and have the sex according to arguments
`ind_sex_permutation` and `nn_sex_permutation`.
}
\details{
Then, a second (random) row is chosen, and the same checks are done for
Individual (2) and NN (2). Furthermore, it is controlled that NN1 was present at
date of second row, and NN2 present at date of first row.

If these conditions are met, the two NNs are flipped.

This procedure is repeated `flips_per_permutation` times.
}
\examples{

}
