% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-redcolobus-data.R
\name{prep_red_colobus_ind_nn_data}
\alias{prep_red_colobus_ind_nn_data}
\title{Prepare Red Colobus scandata for determination of association indices based on Nearest Neighbours:}
\usage{
prep_red_colobus_ind_nn_data(
  scandata_df,
  presence_df,
  id_sex_df,
  start_date,
  end_date,
  interval_length_month,
  filter_adults = TRUE,
  keep_activity_interactant = FALSE,
  ind_sex_included,
  nn_sex_included
)
}
\arguments{
\item{scandata_df}{Data frame with red colobus scan data}

\item{presence_df}{Dataframe created with `get_presence_rc`}

\item{id_sex_df}{Table with sex of individuals. This is the link table
creates with `create_link_table`}

\item{start_date}{String for start date that can be coerced to date by `as.Date`}

\item{end_date}{String for end date that can be coerced to date by `as.Date`}

\item{interval_length_month}{Lenght of each interval in months.}

\item{filter_adults}{Only keep lines where Individual and NN were adult?}

\item{keep_activity_interactant}{If FALSE, re-sort columns and sort out activity, interactant, and inter_age_sex}

\item{ind_sex_included}{Which indivuals (Male, Female) should be included?}

\item{nn_sex_included}{Which nearest neighbours (Male, Female) should be included?}
}
\description{
Prepare Red Colobus scandata for determination of association indices based on Nearest Neighbours:
}
\examples{

}
