% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-redcolobus-data.R
\name{prep_red_colobus_ind_nn_data_one_period}
\alias{prep_red_colobus_ind_nn_data_one_period}
\title{Prepare Red Colobus scandata for determination of association indices based on Nearest Neighbours:
Trim the data frame from start to end of a single period, and only keep relevant columns}
\usage{
prep_red_colobus_ind_nn_data_one_period(
  scandata_df,
  from_date,
  to_date,
  presence_df,
  id_sex_df,
  filter_adults = TRUE,
  keep_activity_interactant = FALSE
)
}
\arguments{
\item{scandata_df}{Data frame with red colobus scan data}

\item{from_date}{Start and end date of analyzed period of time.
Format should be "yyyy-mm-dd". First and last date will be included!}

\item{to_date}{Start and end date of analyzed period of time.
Format should be "yyyy-mm-dd". First and last date will be included!}

\item{presence_df}{Dataframe created with `get_presence_rc`}

\item{id_sex_df}{Table with sex of individuals. This is the link table
creates with `create_link_table`}

\item{filter_adults}{Only keep lines where Individual and NN were adult?}

\item{keep_activity_interactant}{If FALSE, re-sort columns and sort out activity, interactant, and inter_age_sex}
}
\description{
Prepare Red Colobus scandata for determination of association indices based on Nearest Neighbours:
Trim the data frame from start to end of a single period, and only keep relevant columns
}
\examples{

}
