% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-redcolobus-data.R
\name{determine_rc_presence}
\alias{determine_rc_presence}
\title{Determine presence of adult individuals in the Small Camp Red Colobus group
for the specified period of time}
\usage{
determine_rc_presence(scandata_df, first_month_adult_df, from_date,
  to_date, n_min = 5, details = FALSE)
}
\arguments{
\item{scandata_df}{Data frame with red colobus scan data}

\item{first_month_adult_df}{Dataframe with first-month-adult dates}

\item{from_date, to_date}{Start and end date of analyzed period of time.
Format should be "yyyy-mm-dd"}

\item{n_min}{Minimum number of times an individual has to be recorded
(Individual + NN) to be considered present. If recorded < n_min for 3 month
in row, considered as absent (including the first month considered)}

\item{details}{Provide details for each individual and each month (default FALSE)}
}
\description{
This function uses scan data in combination with data on the
first-adult-month of individuals to determine for each month whether an
individual was 1) adult and 2) present
}
\details{
Default is that individuals are considered present if recorded for at least
\code{n_min} (default 5) times per month for 3 months in a row
}
\examples{


}
