% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-sna-pars.R
\name{get_cluster_info_from_timeperiod_list}
\alias{get_cluster_info_from_timeperiod_list}
\title{Function that uses two algorithms (fast_greedy, infomap) to get info about
clusters from simple ratios.}
\usage{
get_cluster_info_from_timeperiod_list(simple_ratio_list, perm.by.layer = NULL)
}
\arguments{
\item{simple_ratio_list}{List created with `get_simple_ratios()``}

\item{perm.by.layer}{In case not all layers (matrices) of permuted data
should be included. This is the `by` argument of the `seq()` function.}
}
\description{
This function uses a list of the Simple Ratio Association Index (or any other
index) for all included time periods. It then calculates modularity and the
ratio of individuals/estimated groups scaled from 0 to 1
(ratio_clusters_inds_scaled) from the observed data and for matrices of permuted
social networks. It use two algorithms from the `igraph` package:
cluster_infomap and cluster_fast_greedy f
}
